修改bug，大致完成切片

+++ dataset.h
@@ -43,7 +43,7 @@ public:
 public:
         Mesh mesh;
         float surroundBox[6];
-        vector<unsigned short> indices;
+        vector<uint> indices;
         vector<float> vertices;
 
         vector<Vertex> vertexset;
